Add DEBUG_SCREENS, improve CMake build, update PORTING status
- CMakeLists.txt: replace custom -DDEBUG with standard CMAKE_BUILD_TYPE=Debug,
  add -DDEBUG_SCREENS option for debug screen switching
- src/debug.c/h: debug module with screen switching (secret msg, game over, hiscore)
- src/sdl_run.c: conditional debug() call under DEBUG_SCREENS
- PORTING.md: update verification status (114/218)
- README/README.ru: add Project Status section, update counts, screenshot
- Multiple game modules: ASM alignment fixes and label corrections

+++ src/game/hiscore_screen.h
@@ -1,11 +1,13 @@
 #ifndef HISCORE_H
 #define HISCORE_H
 
+#include <stdint.h>
+
 /* ASM: Null_both_HiScore (656) */
 void null_both_hi_score(void);
 /* ASM: Draw_Record_HiScore (908) */
 void draw_record_hi_score(void);
-/* ASM: Update_HiScore (4198) */
-int update_hi_score(void);
+/* ASM: Update_HiScore (4198) — возвращает Y: 0/1=1P record/$FF=2P record */
+uint8_t update_hi_score(void);
 
 #endif // HISCORE_H
